Separate benchmark_common into utils and benchmark_input_stream
